Fix warnings. Add popup about invitation status
Former-commit-id: 81a8fc5867d56c447681e43524ccae58299cd6e3

+++ CammentSDK/Modules/Invitation/Presenter/Components/CMInvitationListPresenter.m
@@ -81,7 +81,7 @@
 - (void)appendUsersList:(NSArray<User *> *)users {
     _items = [_items arrayByAddingObjectsFromArray:users];
     NSMutableArray *indexPathesToAdd = [NSMutableArray new];
-    for (int i = _items.count - users.count; i < _items.count; i++) {
+    for (NSInteger i = _items.count - users.count; i < _items.count; i++) {
         [indexPathesToAdd addObject:[NSIndexPath indexPathForRow:i inSection:0]];
     }
     [_tableNode insertRowsAtIndexPaths:indexPathesToAdd withRowAnimation:UITableViewRowAnimationBottom];
